refactor(transaction-data): move instruction-name hooks off the universal barrel
