+++ app/entities/transaction-data/index.ts
@@ -6,7 +6,7 @@ export { getProgramName } from './lib/get-program-name';
 export { getInstructionSummaries, resolveInstructionNames, resolveNamesFromData } from './lib/instruction-summary';
 export { mergeTransactionMap } from './lib/merge-transaction-map';
 export type { InstructionSummary, NamedInstruction } from './lib/types';
-// The second half of instruction naming; the first is `resolveInstructionNames` / `resolveNamesFromData`
-// above. README.md has the whole flow and the invariants it holds to.
-export { useResolvedInstructionNames, useResolvedSummaryNames } from './model/use-resolved-instruction-names';
+// The second half of instruction naming — `resolveInstructionNames` / `resolveNamesFromData` above —
+// is on `client.ts`, since the hooks cannot join a server caller of this barrel. README.md has the
+// whole flow and the invariants it holds to.
 export type { RawTransaction, TransactionConfig, TransactionWithMeta } from './model/types';
